pyproject.toml
[project]
name = "stackystars"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.1.7",
    "ffmpeg-python>=0.2.0",
    "matplotlib>=3.10.8",
    "scikit-image>=0.25.2",
    "tqdm>=4.67.1",
]

[project.scripts]
stackystars = "stackystars.main:cli"

[build-system]
requires = ["uv_build>=0.8.3,<0.9.0"]
build-backend = "uv_build"
